[#28] Fix result_id in Wyscout interception passes
The Wyscout convertor splits passes with an 'interception' tag in two actions: an interception and a pass. The interception got the result_id of the original combined event, which can be problematic when the pass results in an own goal. If this happens, it seems like the player made two own goals instead of one. With this fix, the interception is always labeled as successfull, regardless of what happens after the pass.

Fixes #28