Handle missing players in the public Wyscout data
Some games (e.g., ID=2576016) in the public Wyscout dataset contain substitutions of players with ID=0. This caused a crash of the PublicWyscoutLoader. These players are now excluded from the output of `players()`. If you want to include them, you first have to correct the raw data manually and replace ID=0 with the correct player ID.

Fixes #276