[PublicWyscout] Implement caching for events.json
In the public Wyscout dataset, all events of a season are stored in a single large JSON file. When loading data for a batch of games, this JSON file is now parsed only once and the result is cached for loading successive game. This significantly speeds up the data loader for loading a large number of games from the same season.

Fixes #277