Fix bodypart_onehot feature
This feature generator now sets "bodypart_head/other" to True for "head", "other" and "head/other". This means that for data providers that distinguish between "head" and "other" (i.e., Opta, StatsPerform and StatsBomb), two of the output columns will be True. The behaviour does not change for data providers (i.e., Wyscout) that do not distinguish between the "head" and "other" body parts.

Fixes #275