Improve PublicWyscoutLoader performance (#493)
In the public Wyscout dataset, all events of a season are stored in a single large JSON file. As parsing this file is very expensive, commit b5adcb1 implemented caching. However, the cache was only used in the "events" method. This commit adds caching to the "players" method too, which uses the events to compute the minutes played field.