fix(wyscout): estimate goalmouth coordinates for shots only
The end coordinates of shots are estimated from tags that indicate where the ball
would end up in goal. These tags are also recorded for save attempts and blocked
shots. However, the end coordinate should only be modified for non-blocked shots.
The end-coordinates of blocked shots should be the location of the block.