fix(opta): apply _fix_unintentional_ball_touches only to passes
The _fix_unintentional_ball_touches method was added to deal with "unintentional
ball touch" events (type_id=61 + outcome=0). These are typically deflections. It
sets the result of an action to "success" if the team manages to keep
possession.

But this should only be applied to passes (i.e., actions for wich the result
depends on keeping possession). For example, the result of a shot should depend
on whether a goal was scored.

Fixes #725